added ThemeProvider and used react context to pass theme props to child components
