+++ src/components/ThemePicker/ThemePicker.tsx
@@ -1,19 +1,10 @@
 'use client';
-import { FC } from 'react';
-import { Theme } from '@/types/themeTypes';
 import Image from 'next/image';
-// import { useTheme } from '@/components/ThemeProvider/ThemeProvider';
 import styles from './ThemePicker.module.scss';
+import { useTheme } from '../ThemeProvider/ThemeProvider';
 
-interface ThemePickerProps {
-	theme: Theme;
-	onThemeChange: (theme: Theme) => void;
-}
-
-const ThemePicker: FC<ThemePickerProps> = ({ theme, onThemeChange }) => {
-	// const { theme, toggleTheme } = useTheme();
-	// update onChange to toggleTheme if using ThemeProvider
-	// onChange={() => toggleTheme('y')}
+const ThemePicker = () => {
+	const { theme, handleThemeChange } = useTheme();
 
 	return (
 		<fieldset className={styles.themePicker}>
@@ -28,7 +19,7 @@ const ThemePicker: FC<ThemePickerProps> = ({ theme, onThemeChange }) => {
 						name='theme'
 						value='m'
 						checked={theme === 'm'}
-						onChange={() => onThemeChange('m')}
+						onChange={() => handleThemeChange('m')}
 					/>
 				</label>
 				<label htmlFor='cyan-theme'>
@@ -40,7 +31,7 @@ const ThemePicker: FC<ThemePickerProps> = ({ theme, onThemeChange }) => {
 						name='theme'
 						value='c'
 						checked={theme === 'c'}
-						onChange={() => onThemeChange('c')}
+						onChange={() => handleThemeChange('c')}
 					/>
 				</label>
 				<label htmlFor='purple-theme'>
@@ -52,7 +43,7 @@ const ThemePicker: FC<ThemePickerProps> = ({ theme, onThemeChange }) => {
 						name='theme'
 						value='p'
 						checked={theme === 'p'}
-						onChange={() => onThemeChange('p')}
+						onChange={() => handleThemeChange('p')}
 					/>
 				</label>
 				<label htmlFor='yellow-theme'>
@@ -64,7 +55,7 @@ const ThemePicker: FC<ThemePickerProps> = ({ theme, onThemeChange }) => {
 						name='theme'
 						value='y'
 						checked={theme === 'y'}
-						onChange={() => onThemeChange('y')}
+						onChange={() => handleThemeChange('y')}
 					/>
 				</label>
 			</div>
